+++ internal/controller/api_to_config.go
@@ -29,7 +29,7 @@ type namedRawConfig struct {
 	configName string
 }
 
-func apiToFRR(resources ClusterResources) (*frr.Config, error) {
+func apiToFRR(resources ClusterResources, alwaysBlock []net.IPNet) (*frr.Config, error) {
 	res := &frr.Config{
 		Routers:     make([]*frr.RouterConfig, 0),
 		BFDProfiles: make([]frr.BFDProfile, 0),
@@ -65,8 +65,9 @@ func apiToFRR(resources ClusterResources) (*frr.Config, error) {
 			}
 		}
 
+		alwaysBlockFRR := alwaysBlockToFRR(alwaysBlock)
 		for _, r := range cfg.Spec.BGP.Routers {
-			routerCfg, err := routerToFRRConfig(r, resources.PasswordSecrets, bfdProfiles)
+			routerCfg, err := routerToFRRConfig(r, alwaysBlockFRR, resources.PasswordSecrets, bfdProfiles)
 			if err != nil {
 				return nil, err
 			}
@@ -93,7 +94,7 @@ func apiToFRR(resources ClusterResources) (*frr.Config, error) {
 	return res, nil
 }
 
-func routerToFRRConfig(r v1beta1.Router, secrets map[string]corev1.Secret, bfdProfiles map[string]*frr.BFDProfile) (*frr.RouterConfig, error) {
+func routerToFRRConfig(r v1beta1.Router, alwaysBlock []frr.IncomingFilter, secrets map[string]corev1.Secret, bfdProfiles map[string]*frr.BFDProfile) (*frr.RouterConfig, error) {
 	res := &frr.RouterConfig{
 		MyASN:        r.ASN,
 		RouterID:     r.ID,
@@ -116,7 +117,7 @@ func routerToFRRConfig(r v1beta1.Router, secrets map[string]corev1.Secret, bfdPr
 	}
 
 	for _, n := range r.Neighbors {
-		frrNeigh, err := neighborToFRR(n, res.IPV4Prefixes, res.IPV6Prefixes, r.VRF, secrets, bfdProfiles)
+		frrNeigh, err := neighborToFRR(n, res.IPV4Prefixes, res.IPV6Prefixes, alwaysBlock, r.VRF, secrets, bfdProfiles)
 		if err != nil {
 			return nil, fmt.Errorf("failed to process neighbor %s for router %d-%s: %w", neighborName(n.ASN, n.Address), r.ASN, r.VRF, err)
 		}
@@ -126,7 +127,7 @@ func routerToFRRConfig(r v1beta1.Router, secrets map[string]corev1.Secret, bfdPr
 	return res, nil
 }
 
-func neighborToFRR(n v1beta1.Neighbor, ipv4Prefixes, ipv6Prefixes []string, routerVRF string, passwordSecrets map[string]corev1.Secret, bfdProfiles map[string]*frr.BFDProfile) (*frr.NeighborConfig, error) {
+func neighborToFRR(n v1beta1.Neighbor, ipv4Prefixes, ipv6Prefixes []string, alwaysBlock []frr.IncomingFilter, routerVRF string, passwordSecrets map[string]corev1.Secret, bfdProfiles map[string]*frr.BFDProfile) (*frr.NeighborConfig, error) {
 	neighborFamily, err := ipfamily.ForAddresses(n.Address)
 	if err != nil {
 		return nil, fmt.Errorf("failed to find ipfamily for %s, %w", n.Address, err)
@@ -143,6 +144,7 @@ func neighborToFRR(n v1beta1.Neighbor, ipv4Prefixes, ipv6Prefixes []string, rout
 		EBGPMultiHop: n.EBGPMultiHop,
 		BFDProfile:   n.BFDProfile,
 		VRFName:      routerVRF,
+		AlwaysBlock:  alwaysBlock,
 	}
 	res.HoldTime, res.KeepaliveTime, err = parseTimers(n.HoldTime, n.KeepaliveTime)
 	if err != nil {
@@ -521,6 +523,20 @@ func joinRawConfigs(raw []namedRawConfig) string {
 		res.WriteString("\n")
 	}
 	return res.String()
+}
+
+func alwaysBlockToFRR(cidrs []net.IPNet) []frr.IncomingFilter {
+	res := make([]frr.IncomingFilter, 0, len(cidrs))
+	for _, c := range cidrs {
+		c := c // to make go sec happy
+		filter := frr.IncomingFilter{IPFamily: ipfamily.ForCIDR(&c), Prefix: c.String()}
+		filter.LE = uint32(32)
+		if filter.IPFamily == ipfamily.IPv6 {
+			filter.LE = uint32(128)
+		}
+		res = append(res, filter)
+	}
+	return res
 }
 
 func parseTimers(ht, ka *v1.Duration) (*uint64, *uint64, error) {