+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # FRRK8s Release Notes
 
+## New Release
+
+### Features
+
+- Support a --always-block parameter. The parameter accepts a list of comma separated cidrs to always block. This is useful to protect well known cidrs such as pods or clusterIPs. ([PR 88](https://github.com/metallb/frr-k8s/pull/88))
+
 ## Release v0.0.4
 
 ### Bug fixes